Move EventbusLightPeerChain into trinity.chains
