--- trinity/plugins/builtin/light_peer_chain_bridge/__init__.py
@@ -1,3 +1,0 @@
-from .light_peer_chain_bridge import (  # noqa: F401
-    EventBusLightPeerChain,
-)
